+++ NEWS.md
@@ -1,6 +1,6 @@
 # CHANGES IN xfun VERSION 0.12
 
-
+- Added a new function `split_lines()`.
 
 # CHANGES IN xfun VERSION 0.11